MA-22759-2 add nvm daemon
ELE doesn't have internal non-volatile memory, it depends on the
nvmd (non-volatile memory daemon) to read and write encrypted and
signed copy of persistent data.

This commit implements the nvmd, the persistent data would be
stored to path "/data/vendor/ele/". Storing the persistent data
to secure storage is to-do.

Change-Id: Iabe79a1596c7ac45395ca404550dd91e2eb0ef73
Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25610
Reviewed-by: Elven Wang <elven.wang@nxp.com>