MA-22782 force to use g2d composition for evk_93 to improve UI performance
Avoid checking the layer type and rotation support on the MPU with PXP.
Currently we use PXP to do composition only on the MPU without GPU. Because the
composition with CPU cost too much effort and the UI will be very slow. So we
use PXP to do composition for all cases even some cases PXP cannot process.

Change-Id: I000c9390066a0e967b34945ee61432cfdea5f68c
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25753
Reviewed-by: Jessie Hao <juan.hao@nxp.com>