MA-22860 Fix exposure gain can not recover issue after taking pictures on imx95
If manually adjust the exposure gain and take a photo, then back
to the preview UI, exposure gain will not be restored,
although the value has indeed been set.

The reason may be that under neo ISP, the exposure gain
needs to be set after streaming to truly take effect.

Set the ExposureGain after streaming to fix the issue.
. set m_exposure_gain to "-1" if under Auto exposure mode,
so that the exposure gain can be adjusted normally when
entering the preview UI from camera list UI.

Change-Id: I5633c95fe942b506372f685a8c0396be6af4da0a
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25893
Reviewed-by: Hui Fang <hui.fang@nxp.com>