MA-22895 Fix the incorrect size of pictures taken issue in yuyv format when zoom in on 8mp
Issue occurs when previewing and zooming in to take a photo, photo taken
is still the original size, which does not match the preview size.

When processJpegBuffer, use resizeBuf to handle zoom in,
and will switch the buffers after handleFrame,
update the srcVirtual value after SwitchImxBuf to fix the issue.

Change-Id: I71a3911e1f40386c63700b592324553360353df9
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25977
Reviewed-by: Hui Fang <hui.fang@nxp.com>