MA-23192-1 fix mpl3115 data conflict issue
When TEMPERATURE and PRESSURE instances attempt to open the same device file
/dev/iio:device0, if one is already in poll operation, which means the file
descriptor is in open state, it will cause the other instance to fail to open
the file.

To fix this issue, consider to share the same file descriptor bewteen TEMPERATURE
and PRESSURE instances. When the first instance opens the file, the file
descriptor is saved into the static member variable. Subsequent instances can
check this static member variable. If it has already been opened, can directly
use the existing file descriptor instead of attempting to open it again.

Change-Id: Ifac8f17f98918ab033226486a05f0469693367fc
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26683
Reviewed-by: Bo Zhang <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27001