MA-23192-3 introduce variable iioDevNumCountMap for buffer control
Since the mpl3115 is a combination of temperature and pressure sensors, when the
temperature sensor starts collecting data, if the pressure sensor is disabled at
this time, it will cause the temperature sensor collection to stop. This is
because the trigger and enable nodes of the two bound Buffers are the same.
Once the trigger and enable of one sensor are removed due to being turned off,
the reading of data from the other sensor will naturally stop.

The current patch introduces a global variable named iioDevNumCountMap and adds a
currentIioDevNum for each directory to ensure that the trigger and enable values
of the buffer in a certain directory are only turned off when the count of the
current directory is 0.

Change-Id: Ifcdb2723304b53d0865efa476d50fbc0c74c444e
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26749
Reviewed-by: Bo Zhang <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27003