MA-23425 audio: Increase the nominal latency from 5 to 10ms
This value is used to calculate the minimumStreamBufferSizeFrames.
It is also used as the audio period time.
The current 5ms is too aggressive, the system
can't reach such a low latency, causing some audio glitches,
especially on sof audio playback.

Increase it to 10ms. Note that, a value larger than 10ms
will disable the FastMixer, see calculateBufferSizeFrames in Module.h
and mNormalFrameCount in Threads.cpp.

Change-Id: Ia27ecbb0f5d2685791b7441493d290f8abd97342
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27612
Reviewed-by: Hui Fang <hui.fang@nxp.com>