MA-23574 audio: Check the period size capability
When playing hi-res audio, the period size may be too
large for the card to accept.

Take audio96k16b8c.wav as an example, the mBufferSizeFrames
is 2048 (round up from 96000 * 16ms), the max period size
shown by tinypcminfo is 16376, which is less than
16384 (2048 * 8c), so 2048 is too large a period size.

Check the max period size of a card and limit it to the
default 1024 if it fails.

Change-Id: I73d5f4e6d9ec379cf93e7f0b1e3c0f78a7a0a697
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27825
Reviewed-by: Hui Fang <hui.fang@nxp.com>