WSW-59649 [BT][AIDL] Fix VendorInterface recovery after initialization failure
When VendorInterface::Initialize() fails and Shutdown() is not called,
internal state is not reset. This causes all subsequent Initialize()
attempts to fail, blocking HAL recovery and requiring a system reboot.

Updated code to ensure VendorInterface::Shutdown() is called immediately
after a failed Initialize() attempt to reset the internal state and allow
reinitialization.

Tested: on 8MQ
Change-Id: Ifbf12681bf55f228678f1b018d00ff6efd20b19f
Signed-off-by: Sarveshwar Bajaj<sarveshwar.bajaj@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/28750
Reviewed-by: Yunjie Jia <yunjie.jia@nxp.com>