MA-23901 Enable Hwc3 hdcp interface
The main logic:

when device bootup, the display will be created. HDCP will be
started. The purpose is to obtain the HDCP version of the
currently connected display so that Widevine can obtain the correct HDCP
version earlier. When hdcp authenticate successfully, HDCP thread will
detect the result and record the current hdcp version and hdcp
capability, then stop hdcp.

When surfaceflinger detect secure layer and display is not secure,
startHdcpNegotiation interface will be invoked, and it will start hdcp.
When playback normal video, hdcp will be stopped.

When starting HDCP and stopping HDCP, the state of HDCP will be modified
and surfaceflinger will be notified.
start hdcp: the hdcp information will be sent to surfaceflinger.
stop hdcp: the fake hdcp infomation will be sent to surfaceflinger. so
that surfaceflinger update the hdcp state, the hdcp state will affect
whether start hdcp next time. If don't sent fakehdcp, hdcp will not
start next time.

Change-Id: I09e12b16346c56bfcb9df5f030422459e2a4bb9f
Signed-off-by: He Zhai <zhai.he@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/29200
Reviewed-by: Bo Zhang <bo.zhang@nxp.com>