MA-22827-1 Regard display port as independent parameter instead of display id
The display port can be set by bootargs, while the display id need to be organized
to map with device client and cannot change logic and order. So separate the
display port and display id to avoid such issue.
Refine the parse function of andoridboot.display_port to support hex parameter.

Change-Id: Icdfaf3fae30a4655459c51bbc8e1ee1198507571
Signed-off-by: Bo Zhang <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/29239
Reviewed-by: Zhai He <zhai.he@nxp.com>