MA-23901-3 Optimize the logic of hdcp
For the primary display, hdcp will be started when bootup.
The framework is notified of HDCP changes only when a display hotplug
event occurs.

For secondary display, hdcp will be started during bootup and it wait for
authing successfully then stop. HDCP will be restarted only when a
secure layer exists. When there is no secure layer, HDCP will be
closed and then the framework will be notified.

Change-Id: Ib7084688119d6ca9c7754acdccacd8d0db273218
Signed-off-by: He Zhai <zhai.he@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/29540
Reviewed-by: Bo Zhang <bo.zhang@nxp.com>