MA-23808-1 Use libyuv to do down-scaling instead of fallback to Client composition
The down-scalling operation is not supported in PXP of imx943 A0 chip. So need
to fallback to client composition when process down-scaling layer. So the refresh
rate will be very low, even in home UI.
Here use libyuv to do down-scaling operation and use PXP to do composition to improve
performance.

Change-Id: I9a00eed5037c87e44e7390e643899c90be9765b5
Signed-off-by: Bo Zhang <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/29551
Reviewed-by: Jessie Hao <juan.hao@nxp.com>